PyCSDL2_RWwrite(): cast signed buf.len to unsigned size_t
Fix the signed-unsigned integer comparisons by casting the Py_ssize_t
buf.len to size_t. There should not be any overflow since it should not
be negative.

+++ src/rwops.h
@@ -735,7 +735,7 @@ PyCSDL2_RWwrite(PyObject *self, PyObject *args, PyObject *kwds)
                                      &num))
         return NULL;
 
-    if (buf.len != size * num) {
+    if ((size_t)buf.len != size * num) {
         PyBuffer_Release(&buf);
         return PyErr_Format(PyExc_BufferError, "Invalid buffer size");
     }
